backend/requirements.txt
# Django
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.10

# Config
python-decouple==3.8

# Queue
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# Geo
geopy==2.4.1

# HTTP
httpx==0.28.1

# Tests
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
faker==33.1.0
